Move ButlerConfig and DeferredDatasetHandle out of core.
No real code changes here; just reorganization and import adjustments.

Vision for core is that it should contain things needed by both
Registry and Datastore; these are used by neither.
