+++ python/lsst/daf/butler/__init__.py
@@ -8,4 +8,6 @@ Data Access Butler
 from .core import *
 from .registry import *
 from .butler import *
+from .butlerConfig import *
+from .deferredDatasetHandle import *
 from .version import *
