+++ python/lsst/daf/butler/core/__init__.py
@@ -5,7 +5,6 @@ Core code for butler.
 # Do not export the utility routines from safeFileIo, utils, and queries.
 
 from .assembler import *
-from .butlerConfig import *
 from .config import *
 from .configSupport import LookupKey
 from .composites import *
@@ -25,5 +24,4 @@ from .regions import *
 from .storageClass import *
 from .storedFileInfo import *
 from .dimensions import *
-from .deferredDatasetHandle import *
 from .repoTransfers import *
